:building_construction: Improve project structure
