+++ src/components/TreeList/index.js
@@ -1,4 +1,3 @@
 import _SearchTree from './SearchTree';
 
-
 export const SearchTree = _SearchTree;
