+++ src/pages/Competitors/Workplace.less
