+++ src/pages/Dashboard/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Monitor from './Monitor';
+
+const index = () => {
+  return (
+    <div>
+      <Monitor />
+    </div>
+  );
+};
+
+export default index;
